requirements.txt
numpy<2.0
streamlit>=1.35.0
haystack-ai>=1.20.0
python-dotenv>=1.0.0
anthropic-haystack>=0.1.0  # Para soporte de Anthropic Claude
openai>=1.26.0  # Para soporte de OpenAI
sentence-transformers>=3.0.0
tqdm>=4.66.0
